Fixed a bug related to refreshing cache with tags are added or deleted from dataset, workfile or workspace.

+++ app/models/workfile_version.rb
@@ -22,7 +22,7 @@ class WorkfileVersion < ActiveRecord::Base
 
 
   def delete_cache
-    if self.id != nil
+    if self.id != nil && current_user != nil
       Chorus.log_debug "-- BEFORE SAVE: Clearing cache for #{self.class.name} with ID = #{self.id} --"
       Rails.cache.delete_matched(/.*\/#{self.class.name}\/#{self.id}-#{(self.updated_at.to_f * 1000).round(0)}/)
     end
